backend/requirements.txt
Flask==3.0.2
Flask-Cors==5.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
PyJWT>=2.0.0,<3.0.0
apispec>=6.0.0,<7.0.0
apispec-webframeworks>=0.5.0,<1.0.0
gunicorn==21.2.0
marshmallow>=3.0.0,<4.0.0
psycopg2-binary==2.9.9
pytest==8.0.2
pytest-env==1.1.3
pytest-flask==1.3.0
python-dotenv==1.0.1
flask_jwt_extended